refactor: move creat link to components
