+++ packages/blocks/src/__internal__/rich-text/link-node/index.ts
@@ -1,4 +1,3 @@
 import './affine-link.js';
 
 export { AffineLink } from './affine-link.js';
-export { createLink } from './create-link.js';
